project/requirements.txt
flask
flask-cors
python-dotenv
openai
langchain
langchain-openai
pydantic
google-generativeai
